+++ src/main/java/com/github/cc3002/finalreality/model/character/Enemy.java
@@ -1,6 +1,5 @@
 package com.github.cc3002.finalreality.model.character;
 
-import com.github.cc3002.finalreality.model.character.player.CharacterClass;
 import java.util.Objects;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.TimeUnit;
@@ -10,7 +9,7 @@ import org.jetbrains.annotations.NotNull;
  * A class that holds all the information of a single enemy of the game.
  *
  * @author Ignacio Slater Muñoz
- * @author <Your name>
+ * @author Camila Labarca
  */
 public class Enemy extends AbstractCharacter {
 
@@ -18,10 +17,10 @@ public class Enemy extends AbstractCharacter {
   private final int damage;
 
   /**
-   * Creates a new enemy with a name, a weight and the queue with the characters ready to
-   * play.
+   * Creates a new enemy with a name, points, defense, a weight, damage and the queue
+   * with the characters ready to play.
    */
-  public Enemy(@NotNull final String name, final int points, final String defense, final int weight, final int damage,
+  public Enemy(@NotNull final String name, final int points, final int defense, final int weight, final int damage,
       @NotNull final BlockingQueue<ICharacter> turnsQueue) {
     super(turnsQueue, name, points, defense);
     this.weight = weight;
@@ -35,6 +34,9 @@ public class Enemy extends AbstractCharacter {
     return weight;
   }
 
+  /**
+   * Returns the damage of this enemy.
+   */
   public int getDamage() {
     return damage;
   }
@@ -44,6 +46,10 @@ public class Enemy extends AbstractCharacter {
     super.waitTurn();
     scheduledExecutor
             .schedule(this::addToQueue, this.getWeight() / 10, TimeUnit.SECONDS);
+  }
+  @Override
+  public void addToQueue(){
+    super.addToQueue();
   }
 
   @Override
@@ -55,11 +61,11 @@ public class Enemy extends AbstractCharacter {
       return false;
     }
     final Enemy enemy = (Enemy) o;
-    return getWeight() == enemy.getWeight();
+    return getWeight() == enemy.getWeight() && getDamage() == enemy.getDamage();
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(getWeight());
+    return Objects.hash(getWeight(), Objects.hash(getDamage()));
   }
 }