+++ src/main/java/com/github/cc3002/finalreality/model/character/ICharacter.java
@@ -1,14 +1,12 @@
 package com.github.cc3002.finalreality.model.character;
 
-import com.github.cc3002.finalreality.model.character.player.CharacterClass;
-import com.github.cc3002.finalreality.model.weapon.Weapon;
 
 /**
  * This represents a character from the game.
  * A character can be controlled by the player or by the CPU (an enemy).
  *
  * @author Ignacio Slater Muñoz.
- * @author <Your name>
+ * @author Camila Labarca
  */
 public interface ICharacter {
 
@@ -18,6 +16,9 @@ public interface ICharacter {
    */
   void waitTurn();
 
+  /**
+   * Adds a character to the list {@code turnsQueue}.
+   */
   void addToQueue();
 
   /**
@@ -25,8 +26,14 @@ public interface ICharacter {
    */
   String getName();
 
+  /**
+   * Returns this character's points.
+   */
   int getPoints();
 
-  String getDefense();
+  /**
+   * Returns this character defense.
+   */
+  int getDefense();
 
 }